Refactor ecdsaVerify with EIP712 message creation
